Fix gcc
