+++ conda/README.md
@@ -1,8 +1,12 @@
 ====
-conda install conda-build
-conda install anaconda-client
-anaconda login
+Create conda packages
 
-conda create -n ut python=3.7
-source activate ut
-conda config --add channels conda-forge
+```shell
+$ conda install conda-build
+$ conda install anaconda-client
+$ anaconda login
+$ 
+$ conda create -n ut python=3.7
+$ source activate ut
+$ conda config --add channels conda-forge
+```
