+++ examples/gcc-flags/.gitignore
@@ -1,0 +1,3 @@
+tmp.bin
+cc_flags.json
+cc_params.json
