+++ examples/gcc-flags/apps/matrixmultiply.cpp
