+++ examples/gcc-flags/apps/raytracer.cpp
