+++ examples/gcc-flags/figure.py
