--- examples/gccflags/.gitignore
@@ -1,5 +1,0 @@
-tmp.bin
-cc_flags.json
-gccflags_final_config.cmd
-gccflags_final_config.json
-cc_params.json
