--- examples/gccflags/apps/.gitignore
@@ -1,1 +1,0 @@
-fft.c
