--- examples/heterocl/conv/conv.h
@@ -1,14 +1,0 @@
-#ifndef CONV
-#define CONV
-
-#include <ap_int.h>
-#include <ap_fixed.h>
-#include "typedefs.h"
-
-// conv layers generated by hcl code generator
-void conv_hcl(bit input[MAX_FMAP], const bit W_conv[MAX_W_CONV], const bit8_t threshold[MAX_FMAP], bit output[MAX_FMAP]);
-
-// 2nd conv layers generated by hcl
-void conv_hcl2(bit input[MAX_FMAP], const bit W_conv[MAX_W_CONV], const bit8_t threshold[MAX_FMAP], bit output[MAX_FMAP]);
-
-#endif
