--- examples/heterocl/conv/data/packed_weight_12b
@@ -1,1 +1,0 @@
-1970616514,2265896628,1601672865,2325557993,2986575104,4145902793,1967261635,3592366146,4290866021,3063360788,862205632,2863894510,4117446149,4283467916,3448838761,2653479084,3613864690,542288652,2187577710,2605818413,351988914,4019631152,485584970,4273352796,4034792387,3762437982,2791245578,2014238963,2007134492,2507000557,759738073,4109939720,1516849196,3240496607,2574843042,2242306237,4067579495,2183638706,1225055885,2300045490,1914933280,3602005450,3584965052,912587678,2945767311,963988007,1631994101,2997894094,929365371,3953161679,1081708941,1707282053,3368992309,4117139255,991236256,4248962684,1869023416,3422031744,3996759472,2556182690,907455956,333964285,2475953240,3876711649,600318867,1851648861,2280018805,3770260205,1322943986,1966947323,1968302510,2940357687,2711232988,1423963597,975477050,1532845295,979923893,2980252877,2866108907,2114424319
