--- examples/heterocl/conv/data/packed_weight_5b
@@ -1,1 +1,0 @@
-7931,25078,24051,41736,33544,41752,20046,15912,56847,40490,55584,19124,12381,56454,58002,24317,9024,30537,15883,49638,9110,48793,32789,38155,48874,24736,56443,52779,235,65003,58356,41906,41906,37921,56361,56365,59529,4892,15981,36341,52725,8581,29322,65015,64895,55797,40478,32382,12424,21999,34999,4175,23014,58268,50419,41410,56703,37640,13993,52706,56431,56677,36374,55620,26252,21757,20780,7371,20628,49452,7640,61943,56397,57134,56685,52726,11563,31152,41222,63737,53750,44635,57122,56673,41968,41778,41970,30477,15881,24109,47136,22253,57199,48207,15807,43507,55503,45796,45688,60499,27093,11772,13192,41803,55226,40463,24076,54335,35849,5800,23799,58180,11928,45704,49270,52543,56367,53037,23834,23679,36265,32522,43930,51542,52605,40780,18973,8028,15116,39989,7272,5976,59586,56631,56439,53038,13832,65385,57500,32478,15528,16871,52471,51246,41381,12869,30832,7533,56805,41870,15015,49398,56829,12064,14261,24413,28461,59282,59890,43596,48904,48937,31803,36644,46861,15569,57847,61287,24447,59895,32125,26217,32764,23742,19854,42947,62025,38186,5163,24049,24108,52343,58245,19511,17286,24045,49334,32393,31990,7917,23789,30197,20292,43824,43984,36355,7692,15915,27538,25077,52269,39869,43926,17138,54383,5135,11035,21258,62245,63053,10231,28103,25079,32632,63765,58266,9112,41800,24305,32299,24109,47510,54653,64943,32652,13945,48326,26501,38294,50102,44022,22551,7949,39983,24181,7038,7756,6598,20552,7405,43030,51362,51314,19994,41888,56565,20541,16027,60914,38324,63181,23177,59607,58248,32520,15530,23661,4124,64010,55322,21237,40231,52639,52503,44002,58265,55543,40237,39982,56823,21746,9173,60146,4936,5640,42760,51996,24351,40751,25044,1861,60659,15948,40355,41878,56635,44342,8045
