--- examples/heterocl/conv/data/weight_12b
@@ -1,1 +1,0 @@
-0,0,1 ,1 ,0,1 ,1 ,0,1 ,0,1 ,0,0,0,0,0,1 ,1 ,0,0,0,1 ,0,0,0,0,0,0,1 ,1 ,0,0,0,1 ,0,1 ,0,0,0,0,0,1 ,0,0,0,0,0,0,0,1 ,0,1 ,1 ,1 ,0,0,0,0,1 ,0,1 ,0,0,1 ,0,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,0,1 ,1 ,0,0,0,0,0,0,0,1 ,0,1 ,0,1 ,1 ,0,1 ,1 ,1 ,0,0,1 ,0,1 ,0,0,1 ,1 ,0,1 ,1 ,1 ,0,0,1 ,1 ,0,0,0,0,0,0,1 ,1 ,1 ,1 ,1 ,1 ,1 ,0,1 ,0,0,0,0,1 ,0,0,1 ,0,1 ,0,1 ,1 ,1 ,0,1 ,0,0,0,1 ,0,0,1 ,0,0,1 ,1 ,0,1 ,0,0,0,0,0,1 ,0,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,0,1 ,1 ,0,1 ,0,1 ,1 ,0,1 ,1 ,1 ,1 ,1 ,0,1 ,0,1 ,0,1 ,0,1 ,0,1 ,0,1 ,0,1 ,0,0,1 ,0,1 ,1 ,0,1 ,0,1 ,0,1 ,1 ,0,1 ,0,0,0,0,0,0,0,1 ,0,1 ,0,0,0,1 ,0,1 ,0,0,0,0,1 ,0,0,0,0,1 ,1 ,1 ,1 ,1 ,0,0,1 ,1 ,0,1 ,0,0,1 ,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,1 ,1 ,0,0,1 ,1 ,1 ,1 ,1 ,0,0,1 ,1 ,0,0,0,0,1 ,0,1 ,0,1 ,0,1 ,1 ,1 ,1 ,1 ,1 ,1 ,0,0,0,1 ,1 ,1 ,0,1 ,1 ,1 ,0,1 ,0,0,1 ,1 ,1 ,1 ,0,0,1 ,0,1 ,1 ,1 ,0,1 ,1 ,1 ,0,0,1 ,0,1 ,0,0,0,0,1 ,0,1 ,0,0,0,0,1 ,0,1 ,0,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,1 ,0,1 ,1 ,1 ,0,0,0,0,0,0,1 ,0,0,0,0,1 ,0,0,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,0,0,1 ,0,1 ,0,1 ,0,1 ,1 ,0,1 ,0,1 ,1 ,0,0,0,0,1 ,0,0,0,0,0,1 ,1 ,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,0,1 ,1 ,1 ,1 ,0,0,0,0,0,0,0,0,1 ,1 ,1 ,1 ,0,0,1 ,1 ,0,1 ,0,1 ,0,1 ,0,1 ,0,0,0,1 ,0,1 ,0,0,1 ,1 ,1 ,0,1 ,0,0,0,0,0,1 ,0,0,1 ,0,0,1 ,0,1 ,0,1 ,0,1 ,1 ,1 ,0,1 ,1 ,0,1 ,0,1 ,0,1 ,1 ,0,1 ,1 ,0,0,0,1 ,1 ,1 ,0,1 ,0,0,0,0,0,1 ,0,0,0,0,0,1 ,0,0,1 ,0,0,0,0,0,1 ,0,1 ,1 ,0,0,1 ,0,1 ,1 ,1 ,1 ,0,0,1 ,1 ,0,1 ,0,1 ,0,1 ,1 ,0,0,1 ,1 ,1 ,1 ,0,1 ,0,0,1 ,0,0,0,0,0,1 ,0,1 ,1 ,1 ,0,1 ,0,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,0,1 ,1 ,0,0,1 ,1 ,1 ,1 ,1 ,0,0,0,0,1 ,0,1 ,1 ,1 ,0,0,0,1 ,1 ,0,1 ,1 ,0,1 ,1 ,0,0,1 ,1 ,1 ,1 ,1 ,0,0,0,1 ,0,0,0,0,1 ,1 ,1 ,0,1 ,0,0,0,0,1 ,1 ,1 ,1 ,1 ,1 ,0,1 ,1 ,0,0,0,0,1 ,0,0,1 ,0,1 ,1 ,0,1 ,0,1 ,1 ,1 ,1 ,0,0,0,0,0,1 ,0,1 ,0,0,1 ,1 ,0,0,1 ,1 ,0,1 ,1 ,0,1 ,1 ,1 ,1 ,0,1 ,0,1 ,0,1 ,1 ,0,1 ,1 ,0,0,0,0,0,1 ,0,1 ,0,0,1 ,1 ,1 ,1 ,0,1 ,0,1 ,1 ,0,0,0,1 ,0,0,1 ,0,1 ,0,0,0,0,1 ,1 ,1 ,0,1 ,0,0,0,0,0,1 ,1 ,1 ,0,0,1 ,1 ,1 ,0,0,0,1 ,1 ,0,0,0,1 ,0,0,0,1 ,1 ,0,1 ,0,1 ,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,0,1 ,1 ,1 ,1 ,0,1 ,0,0,1 ,0,1 ,0,1 ,0,0,0,0,1 ,1 ,1 ,1 ,1 ,0,0,0,0,0,1 ,0,1 ,0,0,1 ,0,0,0,1 ,0,0,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,1 ,1 ,0,0,1 ,0,1 ,1 ,0,1 ,1 ,0,1 ,1 ,0,0,0,1 ,0,1 ,1 ,0,1 ,1 ,1 ,1 ,1 ,1 ,0,1 ,0,0,0,1 ,0,1 ,1 ,1 ,0,0,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,1 ,0,0,1 ,1 ,1 ,1 ,1 ,0,0,0,0,1 ,0,0,1 ,0,0,0,0,0,0,1 ,1 ,0,1 ,0,1 ,0,0,1 ,0,1 ,0,1 ,1 ,1 ,1 ,0,0,1 ,0,1 ,1 ,0,1 ,1 ,1 ,0,0,0,1 ,0,0,1 ,0,1 ,1 ,1 ,0,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,1 ,1 ,1 ,1 ,0,0,1 ,0,1 ,1 ,1 ,0,1 ,1 ,0,0,1 ,0,1 ,1 ,1 ,1 ,0,0,1 ,1 ,0,0,1 ,0,1 ,1 ,0,1 ,0,0,0,0,1 ,1 ,0,1 ,1 ,0,1 ,0,0,1 ,1 ,0,1 ,1 ,0,1 ,0,1 ,1 ,1 ,1 ,0,1 ,0,0,1 ,0,1 ,0,0,1 ,0,1 ,0,1 ,0,1 ,0,0,1 ,0,1 ,1 ,1 ,1 ,0,1 ,1 ,0,0,1 ,0,0,1 ,0,0,0,1 ,0,0,1 ,1 ,1 ,0,1 ,1 ,1 ,0,0,0,1 ,0,0,0,0,0,0,1 ,0,0,0,0,0,1 ,1 ,1 ,0,1 ,1 ,0,1 ,1 ,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,1 ,0,0,1 ,1 ,1 ,0,1 ,0,0,1 ,0,1 ,0,1 ,0,1 ,1 ,0,0,1 ,1 ,0,1 ,0,1 ,0,0,1 ,0,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,0,0,1 ,1 ,0,1 ,1 ,1 ,0,0,1 ,0,0,0,0,1 ,0,1 ,0,0,0,1 ,0,1 ,1 ,1 ,0,1 ,0,0,0,1 ,0,1 ,0,1 ,1 ,1 ,1 ,1 ,1 ,0,1 ,0,0,0,0,0,0,1 ,0,0,1 ,0,1 ,0,0,1 ,0,1 ,1 ,1 ,1 ,1 ,1 ,1 ,0,0,1 ,1 ,1 ,0,1 ,0,1 ,0,0,0,0,1 ,0,0,0,1 ,0,0,0,0,1 ,0,0,1 ,1 ,1 ,0,0,1 ,1 ,0,0,1 ,0,0,0,1 ,1 ,0,0,1 ,1 ,0,0,0,1 ,0,0,0,0,1 ,0,1 ,0,1 ,0,0,0,1 ,1 ,0,1 ,1 ,0,1 ,0,1 ,0,0,0,1 ,1 ,0,0,0,0,1 ,0,0,0,0,1 ,1 ,1 ,0,0,1 ,0,1 ,0,0,1 ,0,1 ,1 ,1 ,1 ,1 ,0,1 ,0,0,0,0,1 ,1 ,1 ,0,1 ,1 ,0,1 ,0,1 ,1 ,1 ,0,1 ,1 ,1 ,0,0,1 ,1 ,1 ,1 ,1 ,1 ,0,1 ,0,1 ,1 ,1 ,1 ,0,0,1 ,1 ,1 ,0,0,1 ,1 ,0,0,1 ,0,1 ,0,1 ,0,1 ,0,0,1 ,1 ,0,1 ,0,0,1 ,0,0,0,0,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,0,1 ,1 ,0,1 ,0,1 ,1 ,1 ,1 ,0,0,1 ,1 ,1 ,0,0,0,1 ,1 ,1 ,0,1 ,0,0,0,1 ,0,0,1 ,0,0,0,1 ,1 ,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,0,1 ,1 ,1 ,1 ,1 ,0,0,1 ,1 ,1 ,0,1 ,1 ,1 ,0,0,0,1 ,1 ,1 ,0,0,0,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,0,0,1 ,1 ,0,0,0,1 ,0,1 ,0,0,0,0,1 ,0,0,0,0,1 ,1 ,1 ,0,0,0,1 ,0,0,0,0,0,1 ,1 ,1 ,1 ,1 ,0,1 ,0,0,0,1 ,0,0,0,0,0,0,0,0,1 ,0,0,1 ,1 ,0,1 ,0,0,1 ,1 ,1 ,1 ,0,1 ,0,1 ,1 ,1 ,0,0,0,1 ,0,1 ,1 ,0,1 ,0,0,1 ,1 ,0,1 ,0,0,1 ,0,1 ,0,0,1 ,0,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,0,1 ,1 ,0,0,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,0,0,0,1 ,0,0,0,0,0,1 ,1 ,0,0,0,1 ,1 ,1 ,1 ,1 ,0,1 ,0,1 ,1 ,0,1 ,0,0,1 ,1 ,1 ,1 ,1 ,1 ,1 ,1 ,1 ,1 ,0,0,0,1 ,0,0,1 ,0,1 ,1 ,0,1 ,0,0,1 ,0,1 ,1 ,1 ,1 ,1 ,0,0,0,0,0,0,0,0,0,1 ,0,0,0,0,0,1 ,1 ,1 ,1 ,1 ,0,1 ,0,0,0,0,1 ,0,0,0,0,1 ,1 ,0,0,0,0,0,1 ,1 ,0,1 ,1 ,0,1 ,0,0,0,1 ,1 ,1 ,1 ,1 ,0,1 ,1 ,0,0,0,0,0,1 ,0,0,0,0,1 ,1 ,0,0,1 ,0,1 ,1 ,1 ,1 ,1 ,0,0,1 ,0,0,0,1 ,1 ,0,0,1 ,1 ,1 ,0,1 ,0,1 ,0,0,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,0,0,1 ,1 ,1 ,0,0,1 ,1 ,0,0,1 ,0,1 ,0,1 ,0,0,0,1 ,0,0,0,1 ,0,0,0,0,0,1 ,0,0,0,0,0,0,0,0,0,0,1 ,1 ,0,1 ,1 ,0,0,0,0,1 ,0,1 ,1 ,0,0,1 ,0,1 ,0,1 ,0,0,1 ,1 ,1 ,0,1 ,1 ,0,0,0,0,0,1 ,1 ,1 ,0,1 ,0,0,1 ,1 ,0,1 ,0,0,0,0,1 ,1 ,1 ,1 ,0,0,0,1 ,0,0,1 ,1 ,1 ,1 ,0,0,1 ,0,0,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,0,0,0,0,1 ,0,0,0,1 ,0,1 ,1 ,1 ,1 ,1 ,0,0,0,0,0,0,1 ,0,1 ,1 ,1 ,0,0,1 ,0,1 ,0,1 ,1 ,0,1 ,0,0,0,1 ,0,1 ,1 ,1 ,1 ,1 ,1 ,0,0,0,1 ,1 ,1 ,0,0,0,1 ,0,1 ,0,1 ,1 ,1 ,1 ,0,1 ,0,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,0,0,0,1 ,1 ,0,1 ,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,1 ,1 ,1 ,0,1 ,0,0,0,0,1 ,0,0,0,1 ,1 ,1 ,1 ,0,0,0,1 ,1 ,1 ,0,0,0,1 ,1 ,0,0,1 ,0,1 ,0,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,0,1 ,1 ,0,1 ,0,1 ,0,0,1 ,1 ,1 ,1 ,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,1 ,1 ,1 ,0,1 ,1 ,0,0,1 ,1 ,1 ,1 ,1 ,0,0,1 ,1 ,1 ,0,1 ,1 ,0,1 ,0,0,1 ,1 ,1 ,1 ,0,1 ,1 ,0,1 ,1 ,0,0,0,0,1 ,0,1 ,0,1 ,1 ,0,1 ,0,1 ,0,1 ,0,0,0,0,0,0,1 ,1 ,0,1 ,1 ,1 ,0,1 ,1 ,1 ,0,0,0,1 ,0,1 ,0,1 ,0,1 ,1 ,0,0,0,1 ,1 ,0,1 ,0,1 ,0,1 ,1 ,0,0,1 ,0,1 ,0,1 ,0,1 ,0,1 ,1 ,1 ,0,0,0,0,1 ,0,1 ,0,1 ,1 ,1 ,0,0,1 ,1 ,0,0,1 ,0,0,0,0,1 ,1 ,0,1 ,1 ,1 ,1 ,0,1 ,1 ,0,1 ,0,1 ,1 ,1 ,0,1 ,0,0,0,0,1 ,1 ,0,1 ,0,1 ,0,0,1 ,1 ,1 ,0,1 ,0,0,1 ,1 ,0,1 ,0,1 ,1 ,1 ,0,0,1 ,1 ,1 ,1 ,0,1 ,0,0,1 ,1 ,1 ,1 ,1 ,1 ,1 ,1 ,0,1 ,0,1 ,0,0,1 ,0,1 ,1 ,0,1 ,1 ,0,0,0,0,0,1 ,0,0,1 ,0,1 ,1 ,1 ,0,0,0,0,0,0,1 ,1 ,0,0,1 ,1 ,1 ,0,1 ,0,1 ,0,0,0,1 ,0,1 ,0,1 ,1 ,1 ,0,0,1 ,1 ,0,0,1 ,1 ,0,0,0,1 ,0,1 ,0,1 ,1 ,1 ,1 ,1 ,1 ,1 ,0,0,1 ,1 ,0,0,1 ,1 ,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,0,1 ,1 ,0,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,0,1 ,1 ,0,1 ,0,1 ,0,0,1 ,1 ,1 ,0,0,1 ,1 ,1 ,0,1 ,1 ,0,0,1 ,1 ,0,1 ,0,1 ,1 ,1 ,1 ,0,1 ,1 ,0,1 ,1 ,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,0,1 ,0,1 ,1 ,0,0,0,0,0,0,1 ,0,0,0,0,1 ,0,1 ,0,1 ,0,1 ,0,1 ,1 ,1 ,0,0,1 ,1 ,0,1 ,1 ,1 ,0,0,0,1 ,1 ,1 ,1 ,0,1 ,0,1 ,0,1 ,1 ,0,0,1 ,0,0,0,1 ,1 ,0,1 ,0,1 ,1 ,0,1 ,0,1 ,0,0,1 ,1 ,0,0,0,0,1 ,1 ,0,0,1 ,0,1 ,0,1 ,1 ,1 ,0,1 ,1 ,1 ,0,0,0,1 ,0,1 ,0,0,0,1 ,1 ,1 ,0,0,1 ,1 ,0,0,1 ,1 ,0,1 ,1 ,0,0,0,1 ,0,1 ,1 ,0,1 ,0,0,1 ,0,0,0,0,0,1 ,0,1 ,1 ,0,0,1 ,1 ,0,1 ,0,1 ,1 ,0,1 ,0,0,0,1 ,1 ,0,0,0,1 ,1 ,0,1 ,1 ,1 ,0,0,1 ,0,1 ,0,0,0,1 ,0,0,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,0,1 ,0,0,0,0,0,1 ,0,1 ,0,0,1 ,1 ,1 ,0,1 ,1 ,1 ,0,0,1 ,1 ,1 ,1 ,1 ,0,1 ,1 ,1 ,1 ,0,1 ,0,1 ,1 ,1 ,1 ,1 ,0,0,1 ,0,1 ,0,0,0,1 ,0,1 ,1 ,0,0,0,0,1 ,1 ,0
