--- examples/heterocl/conv/data/weight_13p
@@ -1,1 +1,0 @@
--0.0016707474,0.0197568908,0.0055798823,0.0019061911,-0.0102293976,0.0011146674,-0.0040959688,-0.0043947776,0.0036748489,-0.0039506885
