--- examples/heterocl/env.sh
@@ -1,3 +1,0 @@
-export PYTHONPATH=/work/zhang/common/install/heterocl/python:/work/zhang/common/install/tvm/python/
-source /scratch/xilinx/Vivado_HLS/2017.2/.settings64-Vivado_High_Level_Synthesis.sh
-source /scratch/xilinx/Vivado/2017.2/.settings64-Vivado.sh
