+++ examples/systolic-array/run.tcl
